Resolved the scroll view top space to be displayed without a spacer, and also some small refactoring was done.
